Mark FreeBSD code as obsolete. Pending removal.
FreeBSD is not supported by .NET Core, you're on your own.
There are no plans to support it for .NET 5.
Many things left to implement.
AaruRemote works as of this commit date.
Code will be removed in Aaru 6.0, can be bring back if there's enough interest and future .NET support.